chore: trigger Release Please for v1.3.4 (#73)
This PR triggers Release Please to create a v1.3.4 release PR.

## Context

Release Please was incorrectly trying to create a v2.0.0 release due to
missing v1.3.3 git tags. The tags have now been created and pushed to
the repository.

## Changes

- Empty commit to trigger Release Please workflow
- After this PR is merged, Release Please should create a PR for v1.3.4
(not v2.0.0)

## Testing

This approach was tested successfully in a fork:
https://github.com/petersalomonsen/near-jsonrpc-client-ts/pull/2

The test showed that with the v1.3.3 tags in place, Release Please
correctly generates v1.3.4.